Changed depthMeter type from Integer to Float, and in schema changed from INT to DECIMAL(6,4) (6,4 means total of 6 digits, and 4 of those digits are after the decimal point)

+++ src/main/java/cf/ac/uk/wrackreport/data/jpa/entities/ReportEntity.java
@@ -31,7 +31,7 @@ public class ReportEntity {
     private short depthCategoryId;
 
     @Column(name = "depth_meters")
-    private Integer depthMeters;
+    private Float depthMeters;
 
     @Column(name = "lat_long")
     private String latLong;
